requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.1
Pillow>=10.0
jinja2>=3.1
python-dateutil>=2.8
tenacity>=8.2
pytesseract>=0.3.10
paddleocr>=2.7; platform_machine != "arm64"
paddlepaddle>=2.6; platform_machine != "arm64"
